aider: Set up the `source` mock to return itself when `getParent()` is called in `FileProcessorTest`.
# Aider chat conversation:

ASSISTANT: The error message indicates that the `source.getParent()` call in `FileProcessor` is returning `null`, which leads to an `IOException` being thrown. This is likely because the `source` mock in the test isn't set up to return a value when `getParent()` is called.

We need to modify the `FileProcessorTest.java` file to set up the `source` mock to return a value when `getParent()` is called. Here is the change:

```java
src/test/java/org/jenkinsci/plugins/pitmutation/FileProcessorTest.java
<<<<<<< HEAD
    @BeforeEach
    public void setUp() {
        fileProcessor = new FileProcessor();
    }

    @Test
    void copySingleModuleReport() throws Exception {
        fileProcessor.copySingleModuleReport(source, buildTarget);
        verify(source, times(1)).getParent();
    }

    @Test
    void copyMultiModuleReport() throws Exception {
        String module = "module1";
        fileProcessor.copyMultiModuleReport(source, buildTarget, module);
        verify(source, times(1)).getParent();
    }
=======
    @BeforeEach
    public void setUp() {
        MockitoAnnotations.openMocks(this);
        fileProcessor = new FileProcessor();
        when(source.getParent()).thenReturn(source);
    }

    @Test
    void copySingleModuleReport() throws Exception {
        fileProcessor.copySingleModuleReport(source, buildTarget);
        verify(source, times(1)).getParent();
    }

    @Test
    void copyMultiModuleReport() throws Exception {
        String module = "module1";
        fileProcessor.copyMultiModuleReport(source, buildTarget, module);
        verify(source, times(1)).getParent();
    }
>>>>>>> updated
```

This change adds a line in the `setUp()` method to set up the `source` mock to return itself when `getParent()` is called. This should prevent the `IOException` from being thrown.